% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert2string.R
\name{expert2string}
\alias{expert2string}
\title{Convert IRIDL "expert mode" code to URL string}
\usage{
expert2string(code)
}
\arguments{
\item{code}{String containing IRIDL "expert mode" code}
}
\description{
Replaces spaces with slashes, and trims leading and trailing whitespace.
}
