% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert2string.R
\name{expert2url}
\alias{expert2url}
\title{Convert IRIDL "expert mode" code directly to URL}
\usage{
expert2url(code, return_type = "table", ...)
}
\arguments{
\item{code}{String containing IRIDL "expert mode" code}

\item{return_type}{Options for returning the result. See "Details".}
}
\description{
Convert an IRIDL code string directly to a URL by replacing spaces with
dashes, prepending the IRIDL base URL, and appending a suffix for the
appropriate type of output (see "Details").
}
\details{
The following types of output are supported:
\itemize{
\item \code{table} -- ASCII table; see \code{\link[=returntype_table]{returntype_table()}}
\item \code{ncdf} -- NetCDF file with filename \code{data.nc}.
}
}
