% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_list2string}
\alias{filter_list2string}
\title{Concatenate list of filters to a single string}
\usage{
filter_list2string(filter_list)
}
\arguments{
\item{filter_list}{Vector/List of character filters}
}
\description{
Concatenate list of filters to a single string
}
\seealso{
\code{\link[=filter_point]{filter_point()}}
}
