% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_point.R
\name{filter_point}
\alias{filter_point}
\alias{make_filter_point}
\title{Convert latitude-longitude pair to IRIDL filter point}
\usage{
filter_point(url_string, lat, lon)

make_filter_point(lat, lon)
}
\arguments{
\item{lat}{Latitude, in decimal degrees}

\item{lon}{Longitude, in decimal degrees}
}
\description{
Convert latitude-longitude pair to IRIDL filter point
}
\section{Functions}{
\itemize{
\item \code{make_filter_point}: Create string from lat/lon pair
}}

