% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_url}
\alias{generate_url}
\title{Generate IRIDL URL for a gven return type}
\usage{
generate_url(url_string, returntype = "ncdf4")
}
\arguments{
\item{url_string}{Properly formatted IRIDL URL. See \code{\link[=expert2url]{expert2url()}}.}

\item{returntype}{Type of return}
}
\description{
Generate IRIDL URL for a gven return type
}
