% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_info.R
\name{get_source_info}
\alias{get_source_info}
\title{Get metadata for a source string}
\usage{
get_source_info(source_string)
}
\arguments{
\item{source_string}{Source string}
}
\value{
tibble of information about source
}
\description{
Get metadata for a source string
}
