% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tabular.R
\name{get_tabular}
\alias{get_tabular}
\title{Retrieve climate data from a source in tabular form}
\usage{
get_tabular(source_string, sites_df, ...)
}
\arguments{
\item{source_string}{An expert mode string or character vector}

\item{sites_df}{A data frame of site information. Must contain columns
\code{site_name}, \code{latitude}, and \code{longitude}}
}
\value{
\code{tibble} containing data to be returned
}
\description{
Retrieve climate data from a source in tabular form
}
