% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{get_worldclim_site}
\alias{get_worldclim_site}
\title{Get worldclim data for a bunch of variables for a given site}
\usage{
get_worldclim_site(latitude, longitude, variables = worldclim_vars$label)
}
\arguments{
\item{latitude}{Latitude coordinate}

\item{longitude}{Longitude coordinate}

\item{variables}{Character vector of variables to extract. See \link{worldclim_vars}.}
}
\value{
\code{tibble} containing all variables
}
\description{
Get worldclim data for a bunch of variables for a given site
}
