% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{get_worldclim_sitevar}
\alias{get_worldclim_sitevar}
\title{Get worldclim data for a single site}
\usage{
get_worldclim_sitevar(latitude, longitude, variable)
}
\arguments{
\item{latitude}{Latitude coordinate}

\item{longitude}{Longitude coordinate}

\item{variable}{Character string indicating which variable to download}
}
\description{
Get worldclim data for a single site
}
