% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{iridl_base_url}
\alias{iridl_base_url}
\title{Base URL for IRIDL}
\format{An object of class \code{character} of length 1.}
\usage{
iridl_base_url
}
\description{
Base URL for IRIDL
}
\keyword{datasets}
