% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncep_var.R
\name{ncep_var}
\alias{ncep_var}
\title{Shortcut for adding NCEP variable}
\usage{
ncep_var(url_string, varcode)
}
\arguments{
\item{url_string}{Properly formatted IRIDL URL. See \code{\link[=expert2url]{expert2url()}}.}

\item{varcode}{Type of variable.}
}
\description{
Quickly select a variable from NCEP (\code{SOURCES .NOAA .NCEP .CPC}).
}
\details{
Currently accepted variable names are:
\itemize{
\item \code{"air_temperature"} (\code{.GHCN_CAMS .gridded .deg0p5 .temp}) -- 0.5 degree
gridded surface temperature
\item \code{"precipitation"} (\code{.PRECL .v1p0 .deg0p5 .rain}) -- 0.5 degree gridded
precipitation
-\code{"soil_moisture"} (\code{.GMSM .w})
}
}
