% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_date.R
\name{process_date}
\alias{process_date}
\title{Convert NetCDF date to POSIX}
\usage{
process_date(nclist, remove_original = TRUE)
}
\arguments{
\item{nclist}{NetCDF output as list (see \code{\link[=read_nc2list]{read_nc2list()}})}

\item{remove_original}{(Logical) Remove original variable from list}
}
\description{
Try to guess the date unit from the unit string, and use \code{lubridate} to
convert accordingly.
}
