% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{read_nc2list}
\alias{read_nc2list}
\title{Read all variables from NetCDF file to list}
\usage{
read_nc2list(filename, dims)
}
\arguments{
\item{filename}{Name of file to read}

\item{dims}{Vector or list of NetCDF dimensions to read}
}
\description{
Read all variables from NetCDF file to list
}
