% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{read_ncvar}
\alias{read_ncvar}
\title{Read single variable from NetCDF file}
\usage{
read_ncvar(nc, varid, get_attributes = TRUE, ...)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either 
 function \code{\link[ncdf4]{nc_open}}
 or function \code{\link[ncdf4]{nc_create}}), indicating what file to read from.}

\item{varid}{What variable to read the data from.  Can be a string with the name
 of the variable or an object of class \code{ncvar4}
 If left unspecified, the function will determine if there
 is only one variable in the file and, if so, read from that.  If left unspecified
 and there are multiple variables in the file, an error is generated.
 This argument can also, optionally, specify the name of a dimension (usually the
 unlimited dimension) in order to read values from a coordinate variable.
 Note this is not usual practice, because the \code{ncdim} object already
 contains all the dimension's values in the field named "vals".  However, it can
 sometimes be faster to turn off this automatic reading of the unlimited dimension's
 values by using \code{nc_open(filename, readunlim=FALSE)}, then read the
 dimension values in later with this function.}

\item{get_attributes}{Add attributes from NetCDF variable to returned value}
}
\description{
Read single variable from NetCDF file
}
