% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Download file from URL}
\usage{
retrieve_data(url_string, destfile = tempfile(), ...)
}
\arguments{
\item{url_string}{Properly formatted IRIDL URL. See \code{\link[=expert2url]{expert2url()}}.}

\item{destfile}{A character string with the name where the downloaded
    file is saved.  Tilde-expansion is performed.}

\item{...}{Additional arguments to \code{\link[utils:download.file]{utils::download.file()}}}
}
\description{
Uses \code{\link[utils:download.file]{utils::download.file()}} to download file from an IRIDL URL, and
returns the file name.
}
