% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returntype.R
\name{returntype_table}
\alias{returntype_table}
\title{Return a two-column table}
\usage{
returntype_table(table_opts = NULL, timevar = "T")
}
\arguments{
\item{timevar}{Variable to be used for time. Default = "T".}
}
\description{
Return a two-column table
}
