% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{worldclim_coords}
\alias{worldclim_coords}
\title{Convert lat lon coordinates to two-part Worldclim grid}
\usage{
worldclim_coords(lat, lon)
}
\arguments{
\item{lat}{Latitude coordinate}

\item{lon}{Longitude coordinate}
}
\value{
Named list
}
\description{
Convert lat lon coordinates to two-part Worldclim grid
}
