% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\docType{data}
\name{worldclim_vars}
\alias{worldclim_vars}
\title{Worldclim variables}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 2 columns.}
\usage{
worldclim_vars
}
\description{
Worldclim variables
}
\keyword{datasets}
